change queries to DB from CRUD-spring-data and streams on native sql queries (w.p.)
