+++ src/main/resources/application.properties
@@ -2,3 +2,13 @@ spring.datasource.url=jdbc:postgresql://localhost/new_database
 spring.datasource.username=postgres
 spring.datasource.password=12345678
 spring.jpa.generate-ddl=true
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+logging.level.org.hibernate=ERROR
+logging.level.root=WARN
+logging.level.org.springframework.web=ERROR
+logging.level.com.spa.statistic=DEBUG
+spring.output.ansi.enabled=ALWAYS
+spring.flyway.schemas=test
